backend/pyproject.toml
[project]
name = "excel-dcf-backend"
version = "1.0.0"
description = "Backend API for Excel DCF Assistant"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pypdf2>=3.0.1",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "openai>=1.10.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.3",
    "aiofiles>=23.2.1",
    "requests>=2.31.0",
    "numpy>=1.24.0",
]

[tool.uv]
dev-dependencies = []
